code/ocr/src/requirements.txt
Flask
pillow
pytesseract
